final stuff
its out now go play lol

+++ src/psx/io.c
@@ -40,7 +40,7 @@ void IO_FindFile(CdlFILE *file, const char *path)
 void IO_SeekFile(CdlFILE *file)
 {
 	//Seek to file position
-	CdControlB(CdlSeekL, (u8*)&file->pos, NULL);
+	CdControlB(CdlSeekL, (u32*)&file->pos, NULL);
 }
 
 IO_Data IO_ReadFile(CdlFILE *file)
@@ -70,7 +70,7 @@ IO_Data IO_AsyncReadFile(CdlFILE *file)
 	}
 	
 	//Read file
-	CdControl(CdlSetloc, (u8*)&file->pos, NULL);
+	CdControl(CdlSetloc, (u32*)&file->pos, NULL);
 	CdRead(sects, buffer, CdlModeSpeed);
 	return buffer;
 }
